Refactor how the repositories folder is organized
