+++ app/controllers/beverage.py
@@ -1,0 +1,6 @@
+from ..repositories.managers.beverage import BeverageManager
+from .base import BaseController
+
+
+class BeverageController(BaseController):
+    manager = BeverageManager
