+++ app/repositories/managers/order.py
@@ -1,0 +1,26 @@
+from typing import List
+
+from ..models.order import Order
+from ..models.ingredient import Ingredient
+from ..models.orderDetail import OrderDetail
+from ..serializers import OrderSerializer
+from ..managers.base import BaseManager
+
+class OrderManager(BaseManager):
+    model = Order
+    serializer = OrderSerializer
+
+    @classmethod
+    def create(cls, order_data: dict, ingredients: List[Ingredient]):
+        new_order = cls.model(**order_data)
+        cls.session.add(new_order)
+        cls.session.flush()
+        cls.session.refresh(new_order)
+        cls.session.add_all((OrderDetail(order_id=new_order._id, ingredient_id=ingredient._id, ingredient_price=ingredient.price)
+                             for ingredient in ingredients))
+        cls.session.commit()
+        return cls.serializer().dump(new_order)
+
+    @classmethod
+    def update(cls):
+        raise NotImplementedError(f'Method not suported for {cls.__name__}')
