+++ app/repositories/managers/size.py
@@ -1,0 +1,7 @@
+from ..models.size import Size
+from ..serializers import SizeSerializer
+from ..managers.base import BaseManager
+
+class SizeManager(BaseManager):
+    model = Size
+    serializer = SizeSerializer
