+++ app/repositories/serializers.py
@@ -1,6 +1,9 @@
 from app.plugins import ma
-from .models import Ingredient, Size, Order, OrderDetail
-
+from .models.orderDetail import OrderDetail
+from .models.ingredient import Ingredient
+from .models.size import Size
+from .models.order import Order
+from .models.beverage import Beverage
 
 class IngredientSerializer(ma.SQLAlchemyAutoSchema):
 
@@ -14,6 +17,12 @@ class SizeSerializer(ma.SQLAlchemyAutoSchema):
 
     class Meta:
         model = Size
+        load_instance = True
+        fields = ('_id', 'name', 'price')
+class BeverageSerializer(ma.SQLAlchemyAutoSchema):
+
+    class Meta:
+        model = Beverage
         load_instance = True
         fields = ('_id', 'name', 'price')
