+++ app/test/conftest.py
@@ -5,7 +5,7 @@ import pytest
 from app import create_app, register_blueprints
 from app.plugins import db, ma
 # flake8: noqa
-from app.repositories.models import Ingredient, Order, OrderDetail, Size
+from app.repositories.database.models.orderDetail import Ingredient, Order, OrderDetail, Size
 
 from .fixtures.ingredient import *
 from .fixtures.order import *
